requirements.txt
africastalking==1.2.9
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alabaster==1.0.0
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
asttokens==3.0.0
attrs==25.3.0
babel==2.17.0
black==25.1.0
boto3==1.38.16
botocore==1.38.16
certifi==2025.4.26
cffi==1.17.1
cfgv==3.4.0
channels==4.2.2
channels_redis==4.2.1
charset-normalizer==3.4.2
click==8.2.0
coverage==7.8.0
crispy-bootstrap5==2025.4
decorator==5.2.1
distlib==0.3.9
dj-database-url==2.3.0
Django==5.1.9
django-allauth==65.8.0
django-axes==8.0.0
django-cacheops==7.2
django-cors-headers==4.7.0
django-crispy-forms==2.4
django-csp==4.0
django-debug-toolbar==5.2.0
django-extensions==4.1
django-filter==25.1
django-honeypot==1.2.1
django-htmx==1.23.0
django-modeltranslation==0.19.14
django-ratelimit==4.1.0
django-redis==5.4.0
django-rosetta==0.10.2
django-storages==1.14.6
djangorestframework==3.16.0
docutils==0.21.2
drf-yasg==1.21.10
executing==2.2.0
factory_boy==3.3.3
Faker==37.3.0
filelock==3.18.0
flake8==7.2.0
frozenlist==1.6.0
funcy==2.0
geographiclib==2.0
geopy==2.4.1
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
identify==2.6.10
idna==3.10
imagesize==1.4.1
inflection==0.5.1
iniconfig==2.1.0
ipython==9.2.0
ipython_pygments_lexers==1.1.1
isort==6.0.1
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.0.1
Markdown==3.8
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mccabe==0.7.0
mnemonic==0.20
msgpack==1.1.0
multidict==6.4.3
mypy==1.15.0
mypy_extensions==1.1.0
nodeenv==1.9.1
packaging==25.0
parso==0.8.4
pathspec==0.12.1
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.8
pluggy==1.6.0
polib==1.2.0
pre_commit==4.2.0
prompt_toolkit==3.0.51
propcache==0.3.1
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pycodestyle==2.13.0
pycparser==2.22
pydantic==2.11.4
pydantic_core==2.33.2
pyflakes==3.3.2
Pygments==2.19.1
pyj==0.0
PyNaCl==1.5.0
pytest==8.3.5
pytest-cov==6.1.1
python-dateutil==2.9.0.post0
python-decouple==3.8
python-dotenv==1.1.0
python-json-logger==3.3.0
python-magic==0.4.27
pytz==2025.2
PyYAML==6.0.2
redis==6.1.0
requests==2.32.3
requests-sse==0.5.1
rest-framework-simplejwt==0.0.2
roman-numerals-py==3.1.0
s3transfer==0.12.0
schema==0.7.7
sentry-sdk==2.28.0
six==1.17.0
sniffio==1.3.1
snowballstemmer==3.0.1
Sphinx==8.2.3
sphinx-rtd-theme==3.0.2
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.3
stack-data==0.6.3
stellar-sdk==12.2.1
toml==0.10.2
traitlets==5.14.3
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.4.0
virtualenv==20.31.2
wcwidth==0.2.13
xdrlib3==0.1.1
yarl==1.20.0
